refactory to partial classes.
